[Rename] dir name
