+++ src/code/platform/electron/browser/global.ts
